hsp3: x64 で OLDDLL 関数が負の値を返すと wait が発生する問題を修正
OLDDLL ($202) 規約の関数は BOOL/int (32bit) を返す。MSVC x64 は
\`mov eax, -1\` でゼロ拡張するため rax = 0x00000000FFFFFFFF となり、
exec_dllcmd 側で int64_t result として受け取ると 4294967295 (正の値)
として解釈されてしまう。

result > 0 の分岐に入ると 0x20000 / 0x10000 ビットを見て wait モード
(RUNMODE_AWAIT) に入るため、関数が単に -1 (失敗) を返しただけで
最大 約 36 分の wait ロックが発生していた。新しいプラグイン追加時に
\"hsp3cl_64 がハングする\" 形で再現していた。

修正: x64 経路で call_extfunc の戻り値を int32_t にキャストして
サインエクステンドし直すことで -1 が int64_t -1 として渡るようにした。

+++ hsp3/win32gui/hsp3extlib.cpp
@@ -667,7 +667,12 @@ static int64_t code_expand_next( char *prmbuf, const STRUCTDAT *st, int index )
 		case STRUCTPRM_SUBID_OLDDLLINIT:
 			// 外部 DLL 関数の呼び出し
 #ifdef HSP64
-			result = call_extfunc(st->proc, (INT_PTR *)prmbuf, st->prmmax);
+			// OLDDLL 関数は BOOL/int (32bit) を返すが、CallFunc64 は rax (64bit)
+			// を返す。MSVC は `mov eax,N` でゼロ拡張するため、-1 を返した場合
+			// rax = 0x00000000FFFFFFFF となり、int64_t では 4294967295 (正の値)。
+			// exec_dllcmd 側で result > 0 と判定されてしまうので、int32 として
+			// サインエクステンドし直す。
+			result = (int32_t)call_extfunc(st->proc, (INT_PTR *)prmbuf, st->prmmax);
 #else
 			result = call_extfunc(st->proc, (INT_PTR *)prmbuf, st->size / sizeof(INT_PTR));
 #endif
